amazing how we survived so long with this: [FIX] proper thread control, threads now stop with the applet instead of becoming ghosts and hanging browser
git-svn-id: https://scm.arca.ime.usp.br/morcego/trunk@164 b160122e-fef4-0310-8144-9a7482f4e04c

+++ src/main/java/br/arca/morcego/Config.java
@@ -179,7 +179,7 @@ public class Config {
 
 		// Position of the camera
 		setValue(cameraDistance, new Integer(500));
-		setValue(cameraDepth, new Integer(150));
+		setValue(cameraDepth, new Integer(250));
 		setValue(minCameraDistance, new Integer(150));
 		setValue(fieldOfView, new Integer(250));
